Add slang-help tool to speed up CI cmdline ref check
Extract initCommandOptions() and getCapabilityNames() into lightweight
source files with minimal dependencies, and build a new slang-help
executable that prints the same help as slangc without requiring the
full compiler to be compiled. Updates check-cmdline-ref.yml to build
and use slang-help instead of the full slangc release build.
