+++ .github/workflows/check-cmdline-ref.yml
@@ -28,7 +28,7 @@ jobs:
           config: release
           build-llvm: false
 
-      - name: Build Slang
+      - name: Build slang-cmdline-help
         run: |
           cmake --preset default --fresh \
             -DSLANG_SLANG_LLVM_FLAVOR=DISABLE \
@@ -38,12 +38,12 @@ jobs:
             -DSLANG_ENABLE_SLANGD=OFF \
             -DSLANG_EXCLUDE_DAWN=ON \
             -DSLANG_EXCLUDE_TINT=ON
-          cmake --workflow --preset release
+          cmake --build --preset release --target slang-help
 
       - name: Generate command line reference
         run: |
           mkdir -p temp
-          "$bin_dir/slangc" -help-style markdown -h > temp/command-line-slangc-reference.md
+          "$bin_dir/slang-help" -help-style markdown > temp/command-line-slangc-reference.md
 
       - name: Compare with existing reference
         id: compare
