+++ CMakeLists.txt
@@ -582,6 +582,7 @@ add_subdirectory(prelude)
 add_subdirectory(source/slang-core-module)
 add_subdirectory(source/slang-glsl-module)
 add_subdirectory(source/slang)
+add_subdirectory(source/slang-help)
 add_subdirectory(source/slangc)
 add_subdirectory(source/standard-modules)
 add_subdirectory(examples)
